memory: a size is logged as the size it is
The diagnostic printed when a file cannot be truncated to the size the
bank asked for read that size with a conversion for an int. The value is
a size, which is wider than an int on the platform this branch is
compiled for, so the line printed whatever the conversion took from the
argument and the number in it was not the number asked for. It is printed
as an unsigned long long, which is what the line above it already does
with the same value.

Commit message generated by Claude (Anthropic).

+++ src/lib/xpost_memory.c
@@ -811,7 +811,8 @@ xpost_memory_file_grow(Xpost_Memory_File *mem,
     {
         if (ftruncate(mem->fd, sz) == -1)
         {
-            XPOST_LOG_ERR("ftruncate(%d, %d) returned -1", mem->fd, sz);
+            XPOST_LOG_ERR("ftruncate(%d, %llu) returned -1", mem->fd,
+                          (unsigned long long)sz);
             XPOST_LOG_ERR("strerror: %s", strerror(errno));
         }
     }
